requirements.txt
openai==2.8.1
dotenv==0.9.9
pandas
transformers
torch
pillow